update product routes
