+++ src/controller/api/product/product.controller.js
@@ -4,23 +4,6 @@ import {
   createProduct
 } from '../../../models/Product/product.model.js';
 
-async function httpReadProducts(req, res) {
-  const { search, minPrice, maxPrice } = req.query;
-
-  const filterObj = {};
-  if (search) filterObj.$text = { $search: search };
-  if (minPrice || maxPrice) filterObj.price = {};
-  if (minPrice) filterObj.price.$gte = minPrice;
-  if (maxPrice) filterObj.price.$lte = maxPrice;
-
-  try {
-    const products = await readProductsByFilter(filterObj);
-    return res.status(200).json(products);
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-}
-
 async function httpCreateProduct(req, res) {
   const newProduct = req.body;
   const { imageNames } = req;
@@ -43,4 +26,4 @@ async function httpCreateProduct(req, res) {
   }
 }
 
-export { httpReadProducts, httpCreateProduct };
+export { httpCreateProduct };
