+++ src/controller/router.js
@@ -7,7 +7,7 @@ import productRoutes from './api/product/product.route.js';
 import signupController from './render/user/signup.controller.js';
 import authController from './render/user/auth.controller.js';
 import myAccountController from './render/user/myAccount.controller.js';
-import productRenderRoutes from './render/product/products.route.js';
+import productRenderRoutes from './render/product/product.route.js';
 import orderRenderRoutes from './render/order/order.route.js';
 
 const router = express.Router();
@@ -31,7 +31,10 @@ router.use('/orders', orderRenderRoutes);
 router.get('/cart', authController(['customer']), (req, res) =>
   res.render('customer/cart.ejs')
 );
-router.get('/', (req, res) => res.redirect('/my-account'));
+router.get('/', authController(), (req, res) => {
+  if (req.user === 'customer') return res.render('customer/home.ejs');
+  return res.redirect('/my-account');
+});
 router.use('/*', (req, res) =>
   res.render('404.ejs', { error: null, redirectToLogin: false })
 );
